add: notification for live wallpaper mode #164
